+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The following enhancements and changes have been made to Atari Dev Studio:
  * Updated **Make** Terminal to create a **Command Window** on Windows rather than rely on the default (which is Powershell)
  * Updated the make call to prefix the filename with .\ (thanks lucienEn)
  * Relabled the the **Atari Compiler** window to **ADS Compiler** and the **atari-dev-studio** Terminal window to **ADS Terminal** for consistency.
+ * Cleanup validation of the 7800GD cart option
+ * Added Learning Center examples for batariBasic
 
 ## 0.11.8